Restructure module for python tests
